feat: extract round and response helpers
