+++ src/agent/implementations/CoTAgent.ts
@@ -71,8 +71,8 @@ export class CoTAgent extends BaseReflectionAgent {
           this.agentSetting.documentTag,
         );
 
-        // Then process output files using the parent class method
-        await super.processOutputFiles(
+        // Then process output files using helper
+        await this.fileHandler.processOutputFiles(
           outputFile,
           currRound,
           outputProcessGroupId,
@@ -82,13 +82,13 @@ export class CoTAgent extends BaseReflectionAgent {
       }
 
       // Finally handle statistics in base class (but pass our group ID)
-      const result = await super.handleOutput(
+      const result = await this.fileHandler.handleOutput(
         stateRound,
         stateGlobal,
         outputFile,
         endTurn,
         currRound,
-        outputProcessGroupId, // The statistics will be a subgroup of our output processing group
+        outputProcessGroupId,
       );
 
       // Only end the processing group if we created it
