+++ src/agent/runtime/index.ts
@@ -2,3 +2,6 @@ export * from './executeAgent';
 export * from './agentLoad';
 export * from './ModelFactory';
 export * from './OutputHandler';
+export * from './helpers/ResponseProcessor';
+export * from './helpers/FileHandler';
+export * from './helpers/RoundManager';
